backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.9.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
pytest==7.4.3
httpx>=0.26
psycopg2-binary==2.9.9
python-dateutil==2.8.2

